DAY9-LangChain-Memory-Tools/MAS-Langchain/requirements.txt
qdrant_client
sentence_transformers
langchain
langchain-google-genai
langchain-community
langchain-core
pandas
google-generativeai
python-dotenv